requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.7
django-bulk-update==2.2.0
django-cors-headers==3.2.0
djangorestframework==3.10.3
idna==2.8
importlib-metadata==0.23
itypes==1.1.0
Jinja2==2.10.3
kombu==4.6.6
MarkupSafe==1.1.1
more-itertools==7.2.0
psycopg2==2.8.4
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
